pyproject.toml
[project]
name = "agent-2-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk[a2a]>=1.12.0",
    "python-dotenv>=1.1.1",
]
